refact(socket): move module to `socket::path`
In preparation to adding back the socket communication module, which
is going to replace varlink for internal client/daemon communication.
